requirements.txt
langchain
langchain-google-genai
faiss-cpu
python-dotenv
langchainhub